Develop

+++ sqlapp-command/src/main/java/com/sqlapp/data/db/command/html/UpdateDictionariesCommand.java
@@ -25,7 +25,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.util.Comparator;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -38,7 +38,12 @@ import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellStyle;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.ss.util.CellRangeAddress;
 
+import com.sqlapp.data.db.command.properties.DirectoryProperty;
+import com.sqlapp.data.db.command.properties.FileTypeProperty;
+import com.sqlapp.data.db.command.properties.OutputDirectoryProperty;
+import com.sqlapp.data.db.command.properties.RemoveOriginalFileProperty;
 import com.sqlapp.data.db.command.util.ExcelCommandUtils;
 import com.sqlapp.data.schemas.AbstractDbObject;
 import com.sqlapp.data.schemas.Catalog;
@@ -46,8 +51,8 @@ import com.sqlapp.data.schemas.Column;
 import com.sqlapp.data.schemas.SchemaProperties;
 import com.sqlapp.data.schemas.properties.NameProperty;
 import com.sqlapp.data.schemas.properties.RemarksProperty;
+import com.sqlapp.data.schemas.rowiterator.DataFormat;
 import com.sqlapp.data.schemas.rowiterator.ExcelUtils;
-import com.sqlapp.data.schemas.rowiterator.WorkbookFileType;
 import com.sqlapp.exceptions.InvalidFileTypeException;
 import com.sqlapp.exceptions.InvalidPropertyException;
 import com.sqlapp.util.CommonUtils;
@@ -55,14 +60,38 @@ import com.sqlapp.util.FileUtils;
 import com.sqlapp.util.LinkedProperties;
 import com.sqlapp.util.file.TextFileWriter;
 
-public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UpdateDictionariesCommand extends AbstractSchemaFileCommand
+		implements DirectoryProperty, OutputDirectoryProperty, RemoveOriginalFileProperty, FileTypeProperty {
+
+	private File directory = new File("./");
+	private File outputDirectory = null;
+
+	private String fileType = "xlsx";
+
+	private List<File> inputFiles = new ArrayList<>();
+
+	private List<File> outputFiles = new ArrayList<>();
+
+	private Map<MenuDefinition, Properties> propertiesMap = CommonUtils.linkedMap();
+
+	private boolean dryRun = false;
 
 	private Predicate<String> withSchema = (o) -> true;
 
 	private boolean outputRemarksAsDisplayName = true;
 
+	private boolean removeOriginalFile = true;
+
 	@Override
 	protected void create(Catalog catalog) {
+		inputFiles.clear();
+		outputFiles.clear();
+		propertiesMap.clear();
 		execute(() -> {
 			for (MenuDefinition menuDefinition : MenuDefinition.values()) {
 				createProperties(catalog, menuDefinition, (obj) -> HtmlUtils.objectFullName(obj));
@@ -82,9 +111,12 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		String filename = menuDefinition.toString().toLowerCase();
 		// Properties properties=new SortedProperties(new
 		// StringComparator(this.keywordsMap));
-		Properties fileProperties = new LinkedProperties();
-		Properties mergeProperties = new LinkedProperties();
-		File file = loadProperties(menuDefinition, fileProperties);
+		Properties fileProperties = new Properties();
+		LinkedProperties mergeProperties = new LinkedProperties();
+		File file = loadProperties(this.getDirectory(), menuDefinition, fileProperties);
+		if (file != null) {
+			inputFiles.add(file);
+		}
 		list.forEach(obj -> {
 			String fullName = this.getFullName(obj, this.getWithSchema().test(filename));
 			putProperty(fileProperties, fullName, obj, mergeProperties);
@@ -118,13 +150,26 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		keys.forEach(k -> {
 			putProperty(fileProperties, k, null, mergeProperties);
 		});
-		if (getDictionaryFileDirectory() != null) {
-			if (!getDictionaryFileDirectory().exists()) {
-				getDictionaryFileDirectory().mkdirs();
-			}
+		propertiesMap.put(menuDefinition, mergeProperties);
+		if (dryRun) {
+			return;
 		}
-		if (file == null) {
-			file = new File(this.getDictionaryFileDirectory(), filename + "." + this.getDictionaryFileType());
+		if (getOutputDirectory() != null) {
+			if (!getOutputDirectory().exists()) {
+				getOutputDirectory().mkdirs();
+			}
+			if (file == null) {
+				file = new File(this.getOutputDirectory(), filename + "." + this.getFileType());
+			}
+		} else {
+			if (getDirectory() != null) {
+				if (!getDirectory().exists()) {
+					getDirectory().mkdirs();
+				}
+				if (file == null) {
+					file = new File(this.getDirectory(), filename + "." + this.getFileType());
+				}
+			}
 		}
 		FileUtils.createParentDirectory(file);
 		writeProperties(file, menuDefinition, mergeProperties);
@@ -135,8 +180,10 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		return column.getTableName() + "." + column.getName();
 	}
 
-	private void writeProperties(File file, MenuDefinition menuDefinition, Properties properties) throws Exception {
-		WorkbookFileType workbookFileType = WorkbookFileType.parse(this.getDictionaryFileType());
+	private void writeProperties(File file, MenuDefinition menuDefinition, LinkedProperties properties)
+			throws Exception {
+		properties.sort(new DictionaryKeyStringComparator());
+		DataFormat workbookFileType = DataFormat.parse(this.getFileType());
 		File outputFile;
 		File tempFile;
 		if (workbookFileType != null) {
@@ -144,20 +191,20 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 					+ "." + workbookFileType.getFileExtension());
 		} else {
 			outputFile = new File(file.getParentFile(),
-					FileUtils.getFileNameWithoutExtension(file.getAbsolutePath()) + "." + this.getDictionaryFileType());
+					FileUtils.getFileNameWithoutExtension(file.getAbsolutePath()) + "." + this.getFileType());
 		}
 		if (workbookFileType != null) {
 			tempFile = File.createTempFile(FileUtils.getFileNameWithoutExtension(file.getAbsolutePath()),
 					"." + workbookFileType.getFileExtension(), file.getParentFile());
 		} else {
 			tempFile = File.createTempFile(FileUtils.getFileNameWithoutExtension(file.getAbsolutePath()),
-					"." + this.getDictionaryFileType(), file.getParentFile());
+					"." + this.getFileType(), file.getParentFile());
 		}
-		if ("properties".equalsIgnoreCase(this.getDictionaryFileType())) {
+		if ("properties".equalsIgnoreCase(this.getFileType())) {
 			try (OutputStream os = new FileOutputStream(tempFile)) {
 				properties.store(os, menuDefinition + " dictionaries");
 			}
-		} else if ("xml".equalsIgnoreCase(this.getDictionaryFileType())) {
+		} else if ("xml".equalsIgnoreCase(this.getFileType())) {
 			try (OutputStream os = new FileOutputStream(tempFile)) {
 				properties.storeToXML(os, menuDefinition + " dictionaries");
 			}
@@ -166,12 +213,17 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		}
 		FileUtils.remove(outputFile);
 		tempFile.renameTo(outputFile);
+		if (file != null) {
+			outputFiles.add(tempFile);
+		}
 		if (!file.equals(outputFile)) {
-			file.delete();
+			if (removeOriginalFile) {
+				file.delete();
+			}
 		}
 	}
 
-	private void writeOtherProperties(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
+	private void writeOtherProperties(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
 			Properties properties) throws Exception {
 		if (workbookFileType.isTextFile() && workbookFileType.isCsv()) {
 			writeAsCsv(workbookFileType, file, menuDefinition, properties);
@@ -186,74 +238,94 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		}
 	}
 
-	private void writeAsCsv(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
+	private void writeAsCsv(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
 			Properties properties) throws Exception {
-		List<MenuDefinition> menuDefinitions = CommonUtils.list(menuDefinition.getNest());
-		int maxNestLebel = getMaxNestLebel(properties);
-		if (maxNestLebel != menuDefinition.getNestLevel()) {
-			menuDefinitions.remove(0);
-		}
+		final List<String> headers = getHeaders(menuDefinition, properties);
 		try (FileOutputStream fos = new FileOutputStream(file);
 				OutputStreamWriter writer = new OutputStreamWriter(fos, getCsvEncoding());
 				BufferedWriter bw = new BufferedWriter(writer);
 				TextFileWriter csvWriter = workbookFileType.createCsvListWriter(bw)) {
-			List<String> headers = menuDefinitions.stream().map(c -> c.toString()).collect(Collectors.toList());
-			for (String keyword : getKeywords()) {
-				headers.add(keyword);
-			}
 			csvWriter.writeHeader(headers.toArray(new String[0]));
-			Set<String> output = CommonUtils.set();
+			final Set<String> duplicateCheck = CommonUtils.set();
 			for (Map.Entry<Object, Object> entry : properties.entrySet()) {
-				String key = entry.getKey().toString();
-				if (output.contains(key)) {
+				final String[] values = createWriteData(properties, entry, headers, duplicateCheck);
+				if (values == null) {
 					continue;
-				}
-				String[] values = new String[headers.size()];
-				String value = (String) entry.getValue();
-				int pos = key.lastIndexOf('.');
-				if (pos < 0) {
-					throw new InvalidPropertyException(key, value);
-				}
-				String suffix = key.substring(pos + 1);
-				String[] args = key.split("\\.");
-				for (int i = 0; i < headers.size(); i++) {
-					values[i] = "";
-					String header = headers.get(i);
-					if (i < headers.size() - 2) {
-						values[i] = getHeaderValue(headers, i, args);
-					} else {
-						if (suffix.equalsIgnoreCase(header)) {
-							values[i] = value;
-							output.add(key);
-						} else {
-							String otherKey = key.substring(0, pos) + "." + header;
-							output.add(otherKey);
-							Object otherValue = properties.getProperty(otherKey);
-							if (otherValue != null) {
-								values[i] = otherValue.toString();
-							}
-						}
-					}
 				}
 				csvWriter.writeRow(values);
 			}
 		}
 	}
 
-	private void writeAsWorkbook(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
-			Properties properties) throws IOException {
-		List<MenuDefinition> menuDefinitions = CommonUtils.list(menuDefinition.getNest());
+	private List<String> getHeaders(MenuDefinition menuDefinition, Properties properties) {
+		final List<MenuDefinition> menuDefinitions = CommonUtils.list(menuDefinition.getNest());
 		int maxNestLebel = getMaxNestLebel(properties);
 		if (maxNestLebel != menuDefinition.getNestLevel()) {
 			menuDefinitions.remove(0);
 		}
+		final List<String> headers = menuDefinitions.stream().map(c -> c.toString()).collect(Collectors.toList());
+		for (String keyword : getKeywords()) {
+			String name = keyword;
+			headers.add(name);
+		}
+		return headers;
+	}
+
+	private String[] createWriteData(Properties properties, Map.Entry<Object, Object> entry, List<String> headers,
+			Set<String> duplicateCheck) {
+		String key = entry.getKey().toString();
+		String[] values = new String[headers.size()];
+		String value = (String) entry.getValue();
+		int pos = key.lastIndexOf('.');
+		if (pos < 0) {
+			throw new InvalidPropertyException(key, value);
+		}
+		String suffix = key.substring(pos + 1);
+		String[] args = key.split("\\.");
+		for (int i = 0; i < headers.size(); i++) {
+			values[i] = "";
+			String header = headers.get(i);
+			if (i < headers.size() - 2) {
+				values[i] = getHeaderValue(headers, i, args);
+			} else {
+				if (suffix.equalsIgnoreCase(header)) {
+					values[i] = value;
+				} else {
+					String otherKey = key.substring(0, pos) + "." + header;
+					Object otherValue = properties.getProperty(otherKey);
+					if (otherValue != null) {
+						values[i] = otherValue.toString();
+					}
+				}
+			}
+		}
+		String text = toCheckString(values);
+		if (duplicateCheck.contains(text)) {
+			return null;
+		}
+		duplicateCheck.add(text);
+		return values;
+	}
+
+	private String toCheckString(String[] args) {
+		StringBuilder builder = new StringBuilder();
+		if (args == null) {
+			return builder.toString();
+		}
+		for (String arg : args) {
+			builder.append("_:");
+			builder.append(arg != null ? arg : "");
+		}
+		return builder.toString();
+	}
+
+	private void writeAsWorkbook(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
+			Properties properties) throws IOException {
+		final List<String> headers = getHeaders(menuDefinition, properties);
 		try (FileOutputStream fos = new FileOutputStream(file)) {
 			Workbook workbook = workbookFileType.createWorkbook();
 			Sheet sheet = workbook.createSheet(menuDefinition.toString());
-			List<String> headers = menuDefinitions.stream().map(c -> c.toString()).collect(Collectors.toList());
-			for (String keyword : getKeywords()) {
-				headers.add(keyword);
-			}
+			sheet.setDisplayGridlines(false);
 			int rowNo = 0;
 			org.apache.poi.ss.usermodel.Row row = ExcelUtils.getOrCreateRow(sheet, rowNo++);
 			int cellNo = 0;
@@ -261,54 +333,34 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 			for (String header : headers) {
 				ExcelCommandUtils.setCellValue(row, cellNo++, header, null, cellStyle);
 			}
-			Set<String> output = CommonUtils.set();
+			final Set<String> duplicateCheck = CommonUtils.set();
 			for (Map.Entry<Object, Object> entry : properties.entrySet()) {
-				String key = entry.getKey().toString();
-				if (output.contains(key)) {
+				final String[] values = createWriteData(properties, entry, headers, duplicateCheck);
+				if (values == null) {
 					continue;
-				}
-				String[] values = new String[headers.size()];
-				String value = (String) entry.getValue();
-				int pos = key.lastIndexOf('.');
-				if (pos < 0) {
-					throw new InvalidPropertyException(key, value);
-				}
-				String suffix = key.substring(pos + 1);
-				String[] args = key.split("\\.");
-				for (int i = 0; i < headers.size(); i++) {
-					values[i] = "";
-					String header = headers.get(i);
-					if (i < headers.size() - 2) {
-						values[i] = getHeaderValue(headers, i, args);
-					} else {
-						if (suffix.equalsIgnoreCase(header)) {
-							values[i] = value;
-							output.add(key);
-						} else {
-							String otherKey = key.substring(0, pos) + "." + header;
-							output.add(otherKey);
-							Object otherValue = properties.getProperty(otherKey);
-							if (otherValue != null) {
-								values[i] = otherValue.toString();
-							}
-						}
-					}
 				}
 				row = ExcelUtils.getOrCreateRow(sheet, rowNo++);
 				cellNo = 0;
 				for (String val : values) {
 					Cell cell = ExcelUtils.getOrCreateCell(row, cellNo++);
-					cell.setCellValue(val);
+					if ("".equals(val)) {
+						cell.setCellValue((String) null);
+					} else {
+						cell.setCellValue(val);
+					}
 				}
 			}
 			for (int i = 0; i < headers.size(); i++) {
 				sheet.autoSizeColumn(i);
 			}
+			CellRangeAddress cellRange = new CellRangeAddress(0, 0, 0, headers.size() - 1);
+			sheet.setAutoFilter(cellRange);
+			sheet.createFreezePane(headers.size() - 2, 1);
 			workbook.write(fos);
 		}
 	}
 
-	private void writeAsJson(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
+	private void writeAsJson(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
 			Properties properties) throws IOException {
 		writeAsText(workbookFileType, file, menuDefinition, properties, (map, bw) -> {
 			String text = this.getJsonConverter().toJsonString(map);
@@ -316,15 +368,16 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 		});
 	}
 
-	private void writeAsYaml(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
+	private void writeAsYaml(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
 			Properties properties) throws IOException {
 		writeAsText(workbookFileType, file, menuDefinition, properties, (map, bw) -> {
 			String text = this.getYamlConverter().toJsonString(map);
+			System.out.println(text);
 			bw.write(text);
 		});
 	}
 
-	private void writeAsText(WorkbookFileType workbookFileType, File file, MenuDefinition menuDefinition,
+	private void writeAsText(DataFormat workbookFileType, File file, MenuDefinition menuDefinition,
 			Properties properties, IOEBiConsumer<Map<String, Object>, BufferedWriter> cons) throws IOException {
 		try (FileOutputStream fos = new FileOutputStream(file);
 				BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(fos, "UTF-8"));) {
@@ -424,65 +477,10 @@ public class UpdateDictionariesCommand extends AbstractSchemaFileCommand {
 						}
 					}
 				}
-				if (value != null) {
+				if (!CommonUtils.isEmpty(value)) {
 					mergeProperties.put(key, value);
 				} else {
 					mergeProperties.put(key, "");
-				}
-			}
-		}
-	}
-
-	static class StringComparator implements Comparator<String> {
-
-		private Map<String, Integer> keywordsMap;
-
-		StringComparator(Map<String, Integer> keywordsMap) {
-			this.keywordsMap = keywordsMap;
-		}
-
-		@Override
-		public int compare(String o1, String o2) {
-			String[] split1 = o1.split("\\.");
-			String[] split2 = o2.split("\\.");
-			if (split1.length > split2.length) {
-				return 1;
-			} else if (split1.length < split2.length) {
-				return -1;
-			}
-			int comp = compareWithoutLast(split1, split2);
-			if (comp != 0) {
-				return comp;
-			}
-			return compareLast(split1, split2);
-		}
-
-		private int compareWithoutLast(String[] split1, String[] split2) {
-			for (int i = 0; i < split1.length - 1; i++) {
-				int comp = split1[i].compareTo(split2[i]);
-				if (comp != 0) {
-					return comp;
-				}
-			}
-			return 0;
-		}
-
-		private int compareLast(String[] split1, String[] split2) {
-			String value1 = split1[split1.length - 1];
-			String value2 = split2[split2.length - 1];
-			Integer int1 = keywordsMap.get(value1);
-			Integer int2 = keywordsMap.get(value2);
-			if (int1 == null) {
-				if (int2 == null) {
-					return value1.compareTo(value2);
-				} else {
-					return 1;
-				}
-			} else {
-				if (int2 == null) {
-					return 0;
-				} else {
-					return int1.compareTo(int2);
 				}
 			}
 		}
